+++ packages/react/src/experimental/Forms/Fields/Select/SelectTopActions.tsx
@@ -1,22 +1,28 @@
+import { OneFilterPicker } from "@/components/OneFilterPicker"
 import { GroupingSelector } from "@/experimental/OneDataCollection/Settings/components/GroupingSelector"
 import {
+  FiltersDefinition,
+  FiltersState,
   GroupingDefinition,
   GroupingState,
   RecordType,
 } from "@/hooks/datasource"
+import { useI18n } from "@/lib/providers/i18n"
 import { F1SearchBox } from "../F1SearchBox"
 
 interface SelectTopActionsProps<
   R extends RecordType = RecordType,
   Grouping extends GroupingDefinition<R> = GroupingDefinition<R>,
+  Filters extends FiltersDefinition = FiltersDefinition,
 > {
   showSearchBox?: boolean
+  filters?: Filters
+  currentFilters: FiltersState<Filters>
+  onFiltersChange: (filters: FiltersState<Filters>) => void
   searchBoxPlaceholder?: string
   onSearchChange: (value: string) => void
   searchValue?: string
   searchInputRef: React.RefObject<HTMLInputElement>
-  onFocus?: () => void
-  onBlur?: () => void
   grouping?: Grouping
   currentGrouping?: GroupingState<R, Grouping>
   onGroupingChange?: (grouping: GroupingState<R, Grouping>) => void
@@ -28,26 +34,37 @@ export const SelectTopActions = <R extends RecordType = RecordType>({
   onSearchChange,
   searchValue,
   searchInputRef,
-  onFocus,
-  onBlur,
   grouping,
   currentGrouping,
   onGroupingChange,
+  filters,
+  currentFilters,
+  onFiltersChange,
 }: SelectTopActionsProps<R>) => {
+  const i18n = useI18n()
+
   if (!showSearchBox) return null
   return (
-    <div className="flex gap-2 px-2 pt-2">
-      <div className="flex-1">
-        <F1SearchBox
-          placeholder={searchBoxPlaceholder}
-          onChange={onSearchChange}
-          clearable
-          value={searchValue}
-          key="search-input"
-          ref={searchInputRef}
-          onBlur={onBlur}
-          onFocus={onFocus}
-        />
+    <div className="flex gap-2 border-0 border-b border-solid border-f1-border-secondary p-2">
+      <div className="flex flex-1 flex-row gap-2">
+        <div className="flex-1">
+          <F1SearchBox
+            placeholder={searchBoxPlaceholder ?? i18n.toc.search}
+            onChange={onSearchChange}
+            value={searchValue}
+            key="search-input"
+            ref={searchInputRef}
+            clearable
+          />
+        </div>
+        {filters && (
+          <OneFilterPicker
+            filters={filters}
+            value={currentFilters}
+            onChange={onFiltersChange}
+            mode="compact"
+          />
+        )}
       </div>
       <GroupingSelector
         hideLabel={true}