+++ packages/react/src/experimental/OneCalendar/granularities/day/index.tsx
@@ -144,6 +144,7 @@ export const dayGranularity: GranularityDefinition = {
         motionDirection={renderProps.motionDirection}
         minDate={minDate ? minDate.from : undefined}
         maxDate={maxDate ? maxDate.to : undefined}
+        compact={renderProps.compact}
       />
     )
   },