+++ packages/react/src/experimental/OneCalendar/granularities/month/index.tsx
@@ -152,6 +152,7 @@ export const monthGranularity: GranularityDefinition = {
         motionDirection={renderProps.motionDirection}
         minDate={minDate ? minDate.from : undefined}
         maxDate={maxDate ? maxDate.to : undefined}
+        compact={renderProps.compact}
       />
     )
   },