+++ packages/react/src/ui/InputField/InputField.tsx
@@ -12,6 +12,7 @@ import {
   forwardRef,
   useEffect,
   useId,
+  useRef,
   useState,
 } from "react"
 import { AppendTag } from "./AppendTag"
@@ -266,16 +267,9 @@ const InputField = forwardRef<HTMLDivElement, InputFieldProps<string>>(
       )
     }
 
-    useEffect(
-      () => {
-        if (localValue === value) {
-          return
-        }
-        setLocalValue(value)
-      },
-      // localValue is a dep because we want to recociliate both values, in some cases the value will not change for example when the parent limits the chars the user can input
-      [value, localValue, emptyValue]
-    )
+    useEffect(() => {
+      setLocalValue(value)
+    }, [value])
 
     const handleChange = (
       value: string | React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
@@ -316,6 +310,49 @@ const InputField = forwardRef<HTMLDivElement, InputFieldProps<string>>(
         onClickPlaceholder?.()
       }
     }
+
+    /**
+     * Detect if the input is being autofilled
+     */
+    const [isAutofilled, setIsAutofilled] = useState(false)
+    const handleAnimationStart = (
+      e: React.AnimationEvent<HTMLInputElement>
+    ) => {
+      if (e.animationName === "autofill") {
+        setIsAutofilled(true)
+      }
+    }
+
+    const intervalRef = useRef<NodeJS.Timeout | null>(null)
+    const inputRef = useRef<HTMLElement>(null)
+
+    useEffect(() => {
+      if (isAutofilled && !intervalRef.current) {
+        intervalRef.current = setInterval(() => {
+          if (inputRef.current) {
+            const stillAutofilled =
+              inputRef.current.matches(":-webkit-autofill") ||
+              inputRef.current.matches(":autofill")
+            if (!stillAutofilled) {
+              setIsAutofilled(false)
+              if (intervalRef.current) {
+                clearInterval(intervalRef.current)
+                intervalRef.current = null
+              }
+            }
+          }
+        }, 100)
+      }
+
+      // Cleanup function to clear interval on unmount or dependency change
+      return () => {
+        if (intervalRef.current) {
+          clearInterval(intervalRef.current)
+          intervalRef.current = null
+        }
+      }
+    }, [isAutofilled])
+    /**********************/
 
     return (
       <div
@@ -382,7 +419,7 @@ const InputField = forwardRef<HTMLDivElement, InputFieldProps<string>>(
               <div
                 className={cn(
                   "pointer-events-none absolute left-2 top-[5px] my-auto h-5 w-5 shrink-0",
-                  size === "md" && "left-3 top-2.5"
+                  size === "md" && "left-3 top-[9px]"
                 )}
               >
                 {icon && (
@@ -403,9 +440,11 @@ const InputField = forwardRef<HTMLDivElement, InputFieldProps<string>>(
                 onChange: handleChange,
                 onBlur: props.onBlur,
                 onFocus: props.onFocus,
+                onAnimationStart: handleAnimationStart,
                 disabled: noEdit,
                 readOnly: readonly,
                 role,
+                ref: inputRef,
                 "aria-controls": ariaControls,
                 "aria-expanded": ariaExpanded,
                 id,
@@ -432,7 +471,10 @@ const InputField = forwardRef<HTMLDivElement, InputFieldProps<string>>(
                   (icon || avatar) && "pl-8",
                   (icon || avatar) && size === "md" && "pl-9",
                   inputElementVariants({ size }),
-                  placeholder && !hidePlaceholder && isEmpty(localValue)
+                  placeholder &&
+                    !hidePlaceholder &&
+                    isEmpty(localValue) &&
+                    !isAutofilled
                     ? "opacity-1"
                     : "opacity-0"
                 )}
@@ -446,7 +488,7 @@ const InputField = forwardRef<HTMLDivElement, InputFieldProps<string>>(
               <div
                 className={cn(
                   "flex h-fit items-center gap-1.5 self-center pr-1",
-                  size === "md" && "pr-1.5 pt-1.5",
+                  size === "md" && "pr-1.5",
                   "relative"
                 )}
               >